+++ packages/babel-plugin/src/keyframes/__tests__/call-expression.test.ts
@@ -925,6 +925,9 @@ describe('keyframes', () => {
             const fadeOut = null;
             const StyledComponent = forwardRef(
               ({ as: C = "div", style: __cmpls, ...__cmplp }, __cmplr) => {
+                if (__cmplp.innerRef) {
+                  throw new Error("Please use 'ref' instead of 'innerRef'.");
+                }
                 return (
                   <CC>
                     <CS>{[_, _2, _3, _4]}</CS>
@@ -954,6 +957,9 @@ describe('keyframes', () => {
             const fadeOut = null;
             const StyledComponent = forwardRef(
               ({ as: C = "div", style: __cmpls, ...__cmplp }, __cmplr) => {
+                if (__cmplp.innerRef) {
+                  throw new Error("Please use 'ref' instead of 'innerRef'.");
+                }
                 return (
                   <CC>
                     <CS>{[_, _2]}</CS>
@@ -991,6 +997,9 @@ describe('keyframes', () => {
             const darken = null;
             const StyledComponent = forwardRef(
               ({ as: C = "div", style: __cmpls, ...__cmplp }, __cmplr) => {
+                if (__cmplp.innerRef) {
+                  throw new Error("Please use 'ref' instead of 'innerRef'.");
+                }
                 return (
                   <CC>
                     <CS>{[_, _2, _3]}</CS>
@@ -1026,6 +1035,9 @@ describe('keyframes', () => {
             const fadeOut = null;
             const StyledComponent = forwardRef(
               ({ as: C = "div", style: __cmpls, ...__cmplp }, __cmplr) => {
+                if (__cmplp.innerRef) {
+                  throw new Error("Please use 'ref' instead of 'innerRef'.");
+                }
                 return (
                   <CC>
                     <CS>{[_, _2, _3, _4]}</CS>
@@ -1055,6 +1067,9 @@ describe('keyframes', () => {
             const fadeOut = null;
             const StyledComponent = forwardRef(
               ({ as: C = "div", style: __cmpls, ...__cmplp }, __cmplr) => {
+                if (__cmplp.innerRef) {
+                  throw new Error("Please use 'ref' instead of 'innerRef'.");
+                }
                 return (
                   <CC>
                     <CS>{[_, _2]}</CS>
@@ -1092,6 +1107,9 @@ describe('keyframes', () => {
             const darken = null;
             const StyledComponent = forwardRef(
               ({ as: C = "div", style: __cmpls, ...__cmplp }, __cmplr) => {
+                if (__cmplp.innerRef) {
+                  throw new Error("Please use 'ref' instead of 'innerRef'.");
+                }
                 return (
                   <CC>
                     <CS>{[_, _2, _3]}</CS>
@@ -1305,6 +1323,9 @@ describe('keyframes', () => {
             const fadeOut = null;
             const StyledComponent = forwardRef(
               ({ as: C = "div", style: __cmpls, ...__cmplp }, __cmplr) => {
+                if (__cmplp.innerRef) {
+                  throw new Error("Please use 'ref' instead of 'innerRef'.");
+                }
                 return (
                   <CC>
                     <CS>{[_, _2, _3, _4]}</CS>
@@ -1334,6 +1355,9 @@ describe('keyframes', () => {
             const fadeOut = null;
             const StyledComponent = forwardRef(
               ({ as: C = "div", style: __cmpls, ...__cmplp }, __cmplr) => {
+                if (__cmplp.innerRef) {
+                  throw new Error("Please use 'ref' instead of 'innerRef'.");
+                }
                 return (
                   <CC>
                     <CS>{[_, _2]}</CS>
@@ -1371,6 +1395,9 @@ describe('keyframes', () => {
             const darken = null;
             const StyledComponent = forwardRef(
               ({ as: C = "div", style: __cmpls, ...__cmplp }, __cmplr) => {
+                if (__cmplp.innerRef) {
+                  throw new Error("Please use 'ref' instead of 'innerRef'.");
+                }
                 return (
                   <CC>
                     <CS>{[_, _2, _3]}</CS>
@@ -1406,6 +1433,9 @@ describe('keyframes', () => {
             const fadeOut = null;
             const StyledComponent = forwardRef(
               ({ as: C = "div", style: __cmpls, ...__cmplp }, __cmplr) => {
+                if (__cmplp.innerRef) {
+                  throw new Error("Please use 'ref' instead of 'innerRef'.");
+                }
                 return (
                   <CC>
                     <CS>{[_, _2, _3, _4]}</CS>
@@ -1435,6 +1465,9 @@ describe('keyframes', () => {
             const fadeOut = null;
             const StyledComponent = forwardRef(
               ({ as: C = "div", style: __cmpls, ...__cmplp }, __cmplr) => {
+                if (__cmplp.innerRef) {
+                  throw new Error("Please use 'ref' instead of 'innerRef'.");
+                }
                 return (
                   <CC>
                     <CS>{[_, _2]}</CS>
@@ -1472,6 +1505,9 @@ describe('keyframes', () => {
             const darken = null;
             const StyledComponent = forwardRef(
               ({ as: C = "div", style: __cmpls, ...__cmplp }, __cmplr) => {
+                if (__cmplp.innerRef) {
+                  throw new Error("Please use 'ref' instead of 'innerRef'.");
+                }
                 return (
                   <CC>
                     <CS>{[_, _2, _3]}</CS>