+++ packages/react/src/xcss-prop/index.ts
@@ -3,19 +3,30 @@ import type * as CSS from 'csstype';
 import { ac } from '../runtime';
 import type { CSSPseudos, CSSProperties } from '../types';
 
-type XCSSItem<TStyleDecl extends keyof CSSProperties> = {
+type MarkAsRequired<T, K extends keyof T> = T & { [P in K]-?: T[P] };
+
+type XCSSItem<TStyleDecl extends keyof CSSProperties, TCompiledTypedProperty> = {
   [Q in keyof CSSProperties]: Q extends TStyleDecl
-    ? CompiledPropertyDeclarationReference | string | number
+    ?
+        | CompiledPropertyDeclarationReference
+        | (Q extends keyof TCompiledTypedProperty ? TCompiledTypedProperty[Q] : CSSProperties[Q])
     : never;
 };
 
 type XCSSPseudos<
   TAllowedProperties extends keyof CSSProperties,
   TAllowedPseudos extends CSSPseudos,
-  TRequiredProperties extends { requiredProperties: TAllowedProperties }
+  TRequiredProperties extends { requiredProperties: TAllowedProperties },
+  TCompiledTypedPseudo
 > = {
   [Q in CSSPseudos]?: Q extends TAllowedPseudos
-    ? MarkAsRequired<XCSSItem<TAllowedProperties>, TRequiredProperties['requiredProperties']>
+    ? MarkAsRequired<
+        XCSSItem<
+          TAllowedProperties,
+          Q extends keyof TCompiledTypedPseudo ? TCompiledTypedPseudo[Q] : object
+        >,
+        TRequiredProperties['requiredProperties']
+      >
     : never;
 };
 
@@ -69,7 +80,7 @@ export type XCSSAllProperties = keyof CSSProperties;
 export type XCSSAllPseudos = CSSPseudos;
 
 /**
- * ## xcss prop
+ * ## XCSSProp
  *
  * Declare styles your component takes with all other styles marked as violations
  * by the TypeScript compiler. There are two primary use cases for xcss prop:
@@ -132,10 +143,24 @@ export type XCSSProp<
     requiredProperties: TAllowedProperties;
     requiredPseudos: TAllowedPseudos;
   } = never
+> = Internal$XCSSProp<TAllowedProperties, TAllowedPseudos, object, object, TRequiredProperties>;
+
+export type Internal$XCSSProp<
+  TAllowedProperties extends keyof CSSProperties,
+  TAllowedPseudos extends CSSPseudos,
+  TCompiledTypedProperty,
+  TCompiledTypedPseudo,
+  TRequiredProperties extends {
+    requiredProperties: TAllowedProperties;
+    requiredPseudos: TAllowedPseudos;
+  }
 > =
-  | (MarkAsRequired<XCSSItem<TAllowedProperties>, TRequiredProperties['requiredProperties']> &
+  | (MarkAsRequired<
+      XCSSItem<TAllowedProperties, TCompiledTypedProperty>,
+      TRequiredProperties['requiredProperties']
+    > &
       MarkAsRequired<
-        XCSSPseudos<TAllowedProperties, TAllowedPseudos, TRequiredProperties>,
+        XCSSPseudos<TAllowedProperties, TAllowedPseudos, TRequiredProperties, TCompiledTypedPseudo>,
         TRequiredProperties['requiredPseudos']
       > &
       BlockedRules)
@@ -143,10 +168,8 @@ export type XCSSProp<
   | null
   | undefined;
 
-type MarkAsRequired<T, K extends keyof T> = T & { [P in K]-?: T[P] };
-
 /**
- * ## cx
+ * ## CX
  *
  * Use in conjunction with the {@link XCSSProp} to concatenate and conditionally apply
  * declared styles. Can only be used with the `cssMap()` and {@link XCSSProp} APIs.