+++ packages/utils/src/jsx.ts
@@ -1,2 +1,0 @@
-// Copied from Babel's babel-plugin-transform-react-jsx
-export const JSX_ANNOTATION_REGEX = /^\s*\*?\s*@jsx\s+([^\s]+)\s*$/m;