+++ packages/babel-plugin/CHANGELOG.md
@@ -1,5 +1,73 @@
 # @compiled/babel-plugin
 
+## 0.26.3
+
+### Patch Changes
+
+- a9509cc0: Compiled runtime is no-longer inserted for non-Compiled xcss prop usage.
+
+## 0.26.1
+
+### Patch Changes
+
+- 9857009f: Introduce new API `createStrictAPI` which returns a strict subset of Compiled APIs augmented by a type definition.
+  This API does not change Compileds build time behavior — merely augmenting
+  the returned API types which enforce:
+
+  - all APIs use object types
+  - property values declared in the type definition must be used (else fallback to defaults)
+  - a strict subset of pseudo states/selectors
+  - unknown properties to be a type violation
+
+  To set up:
+
+  1. Declare the API in a module (either local or in a package):
+
+  ```tsx
+  import { createStrictAPI } from '@compiled/react';
+
+  // ./foo.ts
+  const { css, cssMap, XCSSProp, cx } = createStrictAPI<{
+    color: 'var(--ds-text)';
+    '&:hover': { color: 'var(--ds-text-hover)' };
+  }>();
+
+  // Expose APIs you want to support.
+  export { css, cssMap, XCSSProp, cx };
+  ```
+
+  2. Configure Compiled to pick up this module:
+
+  ```diff
+  // .compiledcssrc
+  {
+  +  "importSources": ["./foo.ts"]
+  }
+  ```
+
+  3. Use the module in your application code:
+
+  ```tsx
+  import { css } from './foo';
+
+  const styles = css({ color: 'var(--ds-text)' });
+
+  <div css={styles} />;
+  ```
+
+## 0.26.0
+
+### Minor Changes
+
+- 9860df38: Added error in development/debug mode when using 'innerRef' instead of 'ref'
+
+### Patch Changes
+
+- 52ea5aba: Fix xcss being incompatible with codebases that use Emotion and Compiled:
+
+  - Add `processXcss` option to `@compiled/babel-plugin`. If `processXcss` is set to false, `xcss` usages will be ignored, and will not be processed as Compiled. (Note that `xcss` is currently implemented in Atlassian Design System using Emotion.) Defaults to `true`.
+  - `css` usages in a file will no longer be processed as Compiled if `xcss` is used in the same file, so long as there is not a `@compiled/react` import specified in that file.
+
 ## 0.25.0
 
 ### Patch Changes